requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.0
python-dotenv==1.0.1
requests==2.31.0
requests-mock==1.11.0
six==1.16.0
urllib3==2.2.0
